refactor: Cards comp and rename
